Move weather models to separate folder
